btd_backend/requirements/base.txt
certifi==2017.11.5
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.8
django-appconf==1.0.2
django-debug-toolbar==1.9.1
django-extensions==1.9.8
django-rest-framework==0.1.0
django-rest-swagger==2.1.2
django-reversion==2.0.8
django-rq==0.9.4
Django-Select2==5.8.10
django-treebeard==4.2.0
djangorestframework==3.7.7
html5lib==0.9999999
idna==2.6
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
MySQL-python==1.2.5
olefile==0.44
openapi-codec==1.3.2
Pillow==4.0.0
pytz==2017.3
redis==2.10.5
requests==2.18.4
rq==0.9.2
simplejson==3.13.2
six==1.10.0
sqlparse==0.2.4
typing==3.6.2
uritemplate==3.0.0
urllib3==1.22
